refactor(contacts): extract multiple selection hooks in same directory
