+++ src/legacy/ui-actions/use-move-contacts-multiple-selection.tsx
@@ -9,22 +9,22 @@ import { useModal, useSnackbar } from '@zextras/carbonio-design-system';
 import { every } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
-import { UIAction } from '../../../actions/types';
-import { isRoot, isTrashed, isWriteAllowed } from '../../../carbonio-ui-commons/helpers/folders';
-import { getFolder } from '../../../carbonio-ui-commons/store/zustand/folder';
-import { Folder } from '../../../carbonio-ui-commons/types/folder';
-import { ContactMoveModal } from '../../../components/modals/contact-move';
-import { ACTION_IDS, TIMEOUTS } from '../../../constants';
-import { ContactOrGroup } from '../../../legacy/types/contact';
-import { evaluateParentIds } from '../../../legacy/utils/helpers';
-import { apiClient } from '../../../network/api-client';
+import { UIAction } from '../../actions/types';
+import { isRoot, isTrashed, isWriteAllowed } from '../../carbonio-ui-commons/helpers/folders';
+import { getFolder } from '../../carbonio-ui-commons/store/zustand/folder';
+import { Folder } from '../../carbonio-ui-commons/types/folder';
+import { ContactMoveModal } from '../../components/modals/contact-move';
+import { ACTION_IDS, TIMEOUTS } from '../../constants';
+import { apiClient } from '../../network/api-client';
+import { ContactOrGroup } from '../types/contact';
+import { evaluateParentIds } from '../utils/helpers';
 
 export type MoveContactsAction = UIAction<
 	{ contacts?: Array<ContactOrGroup>; newParentAddressBook?: Folder },
 	{ contacts?: Array<ContactOrGroup>; newParentAddressBook?: Folder }
 >;
 
-export const useActionMoveMultipleSelectionContacts = (): MoveContactsAction => {
+export const useMoveContactsMultipleSelection = (): MoveContactsAction => {
 	const [t] = useTranslation();
 	const createSnackbar = useSnackbar();
 	const { createModal, closeModal } = useModal();
